requirements.txt
django
braintree
python-dotenv
django-crispy-forms
django-verified-email-field
Pillow
pyyaml
